+++ Adj/Program/Controlflow/Implementation/Store.hs
@@ -1,0 +1,6 @@
+module Adj.Program.Controlflow.Implementation.Store where
+
+import Adj.Algebra.Category (type (-->), type (:*:>))
+import Adj.Auxiliary (type (=!?=))
+
+type Store store = (:*:>) store =!?= (-->) store