+++ CHANGELOG.md
@@ -351,3 +351,4 @@
 * Define `<--||--` operator
 * Define `change` method
 * Define `Tagged` datatype
+* Define `Adj.Program.Controlflow.Implementation.Store` module